requirements_no_jax.txt
control==0.9.4
cvxopt==1.3.2
flax==0.8.0
gymnasium==0.29.1
matplotlib==3.8.4
mujoco==3.1.1
mujoco-py==2.1.2.14
numpy==1.26.2
optax==0.1.7
orbax-checkpoint==0.4.8
pandas==2.1.4
pygame==2.5.2
seaborn==0.13.0
stable-baselines3==2.2.1
sb3-contrib==2.2.1
swig==4.1.1
sympy==1.12
tensorboard==2.15.1
tensorflow-probability==0.23.0
tikzplotlib==0.10.1
tqdm==4.66.1
wheel==0.42.0
setuptools==69.0.2
